Cleaned up ID toString chaining in CreateChild Added ResourceIdGenerator class and refactored ChildHydrator Removed redundant code from ChildTransformer Added FamilyAwareTrait
